Preserve vimrc indentation settings by default.
In the original code, the final indentation would be set to DEFAULT_RESULT
when the indentation could not be determined. Now we rather rely on vimrc
settings: if the result is 'tab', the 'tabstop' option should also be
preserved.

The "augroup" block was also moved to a new function FindIndent(), so it can
be reused (for example, when the user wants to detect the indentation of
pasted code).

Patch by Timothy Lin (https://github.com/lzh9102).